requirements.txt
dataclasses_json
mido[ports-rtmidi]
python-osc
python-rtmidi
pyyaml
stupidartnet
zeroconf
pyside6-helpers @ git+https://github.com/MrFrangipane/pyside6-helpers@main
oscartnetfixtures @ git+https://github.com/MrFrangipane/osc-artnet-fixtures@main
